Converted to CMake
